Initial commit.
UI is temp. bare with me on it. it iwll be polished.

We have two approaches to solve any puzzle. first one is where we find the possibilities for each slot and fill it with a number. when we cannot find any solution with the approach one, we use approach two with introduces a random number and see if that can help us to solve the puzzle.

Approach two takes time, but can solve unsolvable puzzles as well.